Update RadiantGI.fx
Update 2.8

I targeted Diffusion setting to work on the floors. Because having it affect everything was not correct to me. I also made the Debug View a bit cleaner.
With The cleaner Debug you  will see more detail. But, It will not cause the final output to show normal edges even if you see them in the Debug View.
 I also made the De-Noiser go up to 20 now so........ Ya...... At any rate it better now!!!!!