Multi Shader Update
RadiantGI update 1.9.8

Updated Temporal AA too account for differences between areas where more noise is likely to be seen and not. This should allow for a more stable image with varied levels of accumulation. So basically it's Adaptive TAA. This was not the only thing added. The other thing was Adaptive Sampling.  Samples will look at a luminance map of the image and reduce sample automatically in areas where It thinks you will not notice a lower quality GI. This means also if you have a bright room it will use more samples and may reduce performance a little. This is still getting tuned and will be adjusted in such a way where it should not be noticeable. As consequence you should be able to lower the 2nd Level Denoiser down a notch or two.

Resolution scale now also adds MIP levels scaling to lower levels. This seems to help with noise at lower levels.

New option added is "Refracted Irradiance control." This allows you adjust the refracted irradiance that hits objects. Like a snow filled landscape and or large body of water adding to the image.  You would be able to control how much that effects the image now. This is not the same as directional sky color. Please look at the MSPaint Informative drawn image. I is art....................

There are other small and minor changes that are done. But, I doubt people will notice. You may now need to readjust your settings slightly.

BloomingHDR was updated it was made easier to use and now uses better default settings.

eMotion Was removed because it has too many problems and I want to work on newer cooler shaders.

Smart Sharp.fx was simplified slightly and it's noise avoidance seems to be working well in most cases.