Create GloomAO.fx
New Shader!

GloomAO

A new Screen Space Directional Occlusion Shader with built in Temporal Anti-Aliasing. As you know SSDO takes direction into account from ambient light in game, to better approximate global illumination.
This physically plausible ambient occlusion also shares some things from my previous GI shader. So think of it as lessons learned shader. I tried to make it easier to use then my older shader GI
With some hidden settings. The shader right now has some issues that are getting worked on still that may cause issues. But, for now it's ready to be released in it's current state. GloomAO is less accurate
than my own GI Shader. But, It allows for back face GI on close objects. TAA is used in this shader for foliage heavy areas and can be adjust to the image. I do hope that the basic settings are a good starting
point for most games in your library.

The Main issues in this shader right now is the Fade out and things like UI and Options are still WIP.
Performance is a important issue that getting worked on as well and right now it's semi adaptive.

Please note that updates will come to this shader.  Here a picture of its Debug view.