Major Update
Removed the reliance on old Overwatch.fxh

The reason for this is that I will start the process of retiring this repo.

The goal is to remove and Strip things and open it up to people who want to work on the shaders.

But, I need to do this in steps over time.